Log files written for screenshot tests
When files are pulled from the device by TradeFed they are renamed, and
the original file name does not end up in the logs. Log the file name on
the device when it is written so it's easier to (1) verify the files
were written at all, (2) in which directory they were written and (3)
what their original file name was.

The log will only be written when the screenshot comparison did not
pass, to be mindfull about log sizes.

Bug: 230070438
Test: N/A
Change-Id: If10603ef9f246e9bbffc522de37dc3b2763d1b10