Implement a workaround for gradle artifact transform
Gradle 7.5 has a bug where it resolves a transform before it is properly
configured. Unfortunately, we hit this bug on the playground build and
the fix couldn't make it to 7.5.

This CL applies a workaround provided by the gradle team.

This CL also updates the playground gradle version to match androidx.

Bug: n/a
Test: CI