Change the Analyzer's target resolution behavior from overriding to fallback
Currently, Analyzer allows its implementation to overrides the target resolution of the ImageAnalysis. This is not flexible if the app want to customize the target resolution. For example, MLKitAnalyzer sets input the resolution to be 640x480 for better performance, but the app may prioritize quality over performance and requires 1080p. The solution is changing the resolution from overriding to falling back to a default resolution.

Updated MLKitAnalyzer's Javadoc to give more background info.

Relnote: rename ImageAnalysis.Analyzer#getTargetResolutionOverride() to ImageAnalysis.Analyzer#getDefaultTargetResolution(). The behavior is also changed so that the value returned by this method can be overridden by the value of ImageAnalysis#setTargetResolution().

Bug: 235128079
Test: manual test and ./gradlew bOS
Change-Id: If1d36af8df968d06c0192ca985c8764c9fb25231