Send repeat key events when a key is held down
Adds functionality to inject repeat `keyDown` events when a key is
held down for a certain length of time. In line with Android
behaviour, the first repeat event occurs 500ms after the initial
`keyDown`, with subsequent repeats occuring at 50ms intervals. The
repeats happen as an implicit side-effect of `advanceEventTime`.

Bug: 233343446
Test: ./gradlew compose:ui:ui-test:test
      && ./gradlew compose:ui:ui-test:cC
Change-Id: I03c9c10c3fe16a8395ab386cb9951365ed6d5c27