Send preDraw message at front of queue pre API 24
Prior to API level 24, JankStats uses an OnPreDrawListener to track frame
times. The handler for posting these messages originally posted them at the
front of the queue. This CL returns to that logic.

Bug: 233358407
Test: JankStatsTest runs, manual testing
RelNote: "Return to original logic of posting OnPreDrawListener messages at
front of queue, for more consistent and predictable frame timing."

Change-Id: I05a434fe9453ea1be28d398e3eb284dd9b0cb64a