Reverse doc staging allowlist to make dackka the default
There are some libraries that aren't in the tracking spreadsheet:
- textclassifier
- draganddrop
- glance

To test this, I modified the original version of the script based on b/231442534#comment7 to remove the dokka/doclava directories before copying dackka versions over. I ran the script up to the "Create (if needed) and sync g4 workspace" step and copied the output docs to a different directory. I ran the new version of the script up to the same step, and took the diff of the old version of the generated docs and the new version.
The class index files are different in the two versions. I copied the class lists into a spreadsheet to compare them: go/dackka-staging-diff. The "only in new list" column has the classes that are only in the new version, and vice versa for the old list. There are separate tabs for Java and Kotlin. From checking a few, it seems like the differences are companion objects, *Kt classes, and whether type parameters are used.
The only other diff is that the old version has docs for android.support.wearable (not androidx) in Kotlin. As far as I can tell these pages don't exist in the published docs.

Test: described above
Fixes: 231442534
Change-Id: Ia9b239b3b5be79ce2c67cc8692eb39d5257bf682