Fixed double-removal of listener delegator
It was possible for the overall listener delegator object to be
removed twice, in both places where the delegator's list of
delegates could reach zero. This fix ensures that it will only ever
be removed once (when that list first reaches a size of zero).

Also cleaned up historical use of Executor in the JankAggregatorActivity
test app.

Bug: 239457413
Test: Manual testing plus JankStatsTest automated runs
RelNote: Fixed crash due to double-removal of OnFrameMetricsAvailableListener
Change-Id: I4409483d6e2f7287a0a93f521f68a4be9e22d969