Add cross-UID embedding sample
A new module is added to host activities in a different UID that the
primary sample can embed. The primary sample module is signed with a
certificate that is known to a test activity in a new module to test
trusted embedding. The primary sample is moved one level down.

The new module structure is:
- window:window-demos:demo
- window:window-demos:demo-second-app

Bug: 231499394
Test: Build and run :demo-second-app and :demo
Change-Id: I1c98e3f929c8f12a508deee3b26be3bf156fa9ef