Implement PendingIntentCompat to enforce mutability flag.
To improve system security, M+ starts to support the IMMUTABLE, and S+ requires that mutability Flags must be set. On S+. When an app creates a PendingIntent without providing mutability flags, the app will be force-closed during runtime. Choosing to add or not add Mutability Flags depending on the operating system version is a hassle, not to mention adding version check. Moreover, mutability is easy to be forgotten to add. When we separate mutability into a Boolean in PendingIntentCompat, we can force this parameter be set directly at compile time to reduce runtime errors.