Ellipsizing when height allows for 0 lines
When implementing  'ellipsis with height' we skipped the case when calculatedMaxLines was zero (which is when there's no vertical space to fit even a single line). With this change we will remeasure such a case with maxLines = 1 allowing for the first line to be ellipsized. In this case even if there's only a tiny little difference between the maxHeight and height of one line, we can show the ellipsized line.

Also changed how paragraph is drawn for Ellipsis overflow - now it will clip. This doesn't change behaviour for most cases, but it fixes the case when the height allows for less than one line.

Bug: 168720622
Test: new test in AndroidParagraphTest.kt
Test: demo -> Text -> Ellipsize

Change-Id: I92a7a5b92c729cd17bdc307ddb85fa902807339f