Fix issue in `JavaPoetExt.toAnnotationSpec()`
This CL fixes an issue with `JavaPoetExt.toAnnotationSpec()` when
dealing with enums (and potentially other types) where only the simple
name of the enum was output rather than the qualified name.

This CL changes the implementation to use XAnnotationValue's hasser
methods to get instances XEnumTypeElement rather than using
`instanceof` on the value type directly.

This CL also adds tests for every kind of annotation value to ensure
that the AnnotationSpec is created correctly.

This CL also fixes an ordering issue, where KSAnnotation.arguments
wasn't returning the annotation values in the order they were defined in
the java/kotlin source. This ordering is necessary for stability, e.g.
in compiler tests, to prevent getting slightly different AnnotationSpecs
depending on the backend (Javac/KAPT/KSP).

Test: XAnnotationValueTest
Change-Id: Ic4e9e5d5148ea4e4bb8b345c97dbff1169e53060