Remove nullability annotations added in 1.3.0-beta01
These were a source-incompatible change in Kotlin, which caused a lot of
pain for clients attempting to update to 1.3-0-beta01. Since we do not
currently have tooling for a proper deprecate/remove cycle
(see https://issuetracker.google.com/240600673), we are removing the
annotations until the ability to do so exists.

https://issuetracker.google.com/240775049 is also filed to add the
annotations back once that is done.

Relnote: "Remove nullability annotations added in 1.3.0-beta01"

Bug: 238355930
Test: API change, validating via presubmit and post-merge drop to google3
Change-Id: I7a25874d17e40de21c6bd2f50192c79746d6b7e5