Support descriptors with LOCAL nesting type elements.
It is not possible to discover a LOCAL type element compiled from Java sources via javac, but kotlinc does in some occasions produces bytecode that uses a LOCAL type in a method signature. Specifically for private functions with a right-hand side anonymous expression. Meanwhile KAPT stubs do not use the LOCAL type and instead will either use java.lang.Object or the single supertype of the LOCAL type element.

Therefore to support finding KmFunction metadata which uses descriptors, this CL adds support for getting the internal name of a LOCAL nested type element.

Bug: 232742201
Test: KotlinMetadataElementTest
Change-Id: Ia61c22cf562f4f3928f1c33d75c8acb4f8a4566f