Adds experimental WeightedElementsComplicationData
Bug: 240960805
Test: Added new test cases
Relnote: We've added experimental support for WeightedElementsComplicationData which consists of an array of Elements (pairs of weight and color) along with optional text/title/image. These might be displayed as a pie chart where the colors need to be meaningful given the context, since there typically isn't room in a complication to render labels.
Change-Id: I87eea137c8c4cf14921139f78d81a40433169f2e