Move away from caching configurations in buildSrc/shared.gradle
AndroidX has special code to serialize/deserialize configurations in
buildSrc/shared.gradle. We had cacheableImplementation/cacheableApi
extending implementation/api. However, because we resolve down to jars
for cacheableImplementation and cacheableApi separately and put those
jars on the two classpaths, it can lead to issues where a dependency
version is bumped under a library. This is exactly what happens when
trying to update wire plugin (r.android.com/2175906) that depends on
okio in cacheableImplementation configuration, but okio at a different
version is also pulled in by another dep in cacheableApi configuration.

To remove the risk of hitting this issue, remove this caching layer
that is now mostly redundant with Gradle configuration caching enabled.

Also upgrading to wire 4.4.1 as that fixes kotlinpoet classpath issues

Test: ./gradlew :benchmark:benchmark-common:generateDebugProtos success
      with r.android.com/2175906 local patch.
Change-Id: Ib603a54d5970e8e29284a9cef62a994ca0bfdb5c