Add nested indexing support for Documents.
Currently, the indexNestedProperties is ignored when the Document
annotations are parsed and converted into an AppSearch schema. This means
that the only way to index nested documents is to manually generate a
schema, despite the indexNestedProperties property already existing on the
DocumentProperty annotation.

This CL adds support for the SchemaCodeGenerator to add the required index
nested properties function to the schema builder generation, which means
nested documents now can be indexed by AppSearch when using the Document
pattern.

Additionally, there is a new test for checking that the
indexNestedProperties field is respected when generating the code,
including ensuring it has the correct value when omitted entirely.

The golden file for the testToGenericDocument_allSupportedTypes has also
been updated to reflect this change.

Bug: 240842741
Test: AppSearchCompilerTest
Change-Id: Iae9a67a03e8f2d6447578c0c26a53d03290b493f