Add generation to WorkerInfo
Relnote: """
`WorkerInfo.getGeneration()` was added that returns
a generation of a worker. A worker has multiple generations, if it was updated via
`WorkManager.updateWork` or `WorkManager.enqueueUniquePeriodicWork` using
`ExistingPeriodicWorkPolicy.UPDATE`.
If the worker is currently running, it can possibly be of an older generation rather than
returned by this methof if an update has happened during an execution of this worker."""
bug: 219446409
Test: WorkUpdateTest

Change-Id: I665c5d480700dd682d1268ee88be124783fca70f