Create a BundlingNode and a ProcessingNode.
BundlingNode: a node that listens to incoming images and request, then bundle images with request. Add a implementation for the basic case where is one image per request. For extensions where there are multiple images per request, we leave it to the extension to bundle it.

ProcessingNode: a node that runs post-processing jobs on a single image.

Bug: 241803593
Test: manual test and ./gradlew bOS
Change-Id: If5cf4186b4b49fc22f64ff5344b7b33da323e834