CameraSurfaceManager - A class that manages the lifetime of Surfaces
This CL adds CameraSurfaceManager - A new class that manages the
lifetime of Surfaces in CameraPipe.

This class has the same lifetime of CameraPipe, and continuously
monitor the liftime of Surfaces as CameraGraphs are started and
stopped.

When a CameraGraph is started, the initial list of Surfaces along with
the latter Surfaces set by CameraGraph.setSurface() will be registered
at CameraSurfaceManager where each Surface will be given a token. Note
that if the same Surface is used in another CameraGraph, it will be
given another token. Essentially each token means a single use on a
Surface.

Tokens are closed when a capture session ended (i.e., closed, failed),
or when a CameraGraph is stopped.

Bug: 243701641
Test: CameraPipe test app; CameraSurfaceManagerTest
Change-Id: Ifca5331f8df6642fc37be0c3780002df0e10eeb8