Update core test app tests to reset the cameraImpl
For Parameterized test with camera-camera2/camera-pipe,
Tests (using CameraXActivity) will shut down camerax in
the @After block, and it will require to force config
the cameraImpl every time when start the test Activity.

Add an extra Intent key to ignore the configured
implementation type, so it will be able to configure
the camera implement again when next time to start the
Activity.

Bug: 243758509
Test: TakePictureTest, ToggleButtonUITest,
ExistingActivityLifecycleTest

Change-Id: Idf674d9f90145d59a9d7311c6c15d79468571b5a