Fix semantics during delayed parameter extraction
In a recent CL we changed the way we compute parameters. Instead of
walking the slot tree we make a lookup by anchor id.

However the semantics are returned from the result of the
GetParametersCmd, and the GetParameterDetailsCmd. The semantics
are not computed from parameters, they are computed during the
parsing of the slot tree. Thus the semantics were lost when we
started to lookup by anchor id.

Fix this by using the node that was cached from the previous
GetComposablesCmd.

Fixes: 244002984
Fixes: 181811685
Test: Added unit test for semantics using delayed param extraction.
Change-Id: Ia3c0c2cf5a974010f4d0a5ce0b0a546b2343de4e