Fix audio buffer dropped due to same timestamp as previous buffer
The root cause is that AudioRecord#getTimestamp() is not guaranteed to
be synchronized with AudioRecord#read(). AudioRecord#getTimestamp(), and
might return the same timestamp after calling AudioRecord#read().

Track the total number of audio data read and use
AudioRecord#getTimestamp() as a reference to interpolate the timestamp
we should pass to the encoder.

Bug: 244265927
Test: manual test, ./gradlew bOS and ./gradlew camera:camera-video:testDebug
Change-Id: I36cc237c8a9fddf9b9694663d300cc5864670b3f