[GH] Implement PendingIntentCompat to enforce mutability flag.
To improve system security, M+ starts to support the IMMUTABLE, and S+ requires that mutability Flags must be set. On S+. When an app creates a PendingIntent without providing mutability flags, the app will be force-closed during runtime. Choosing to add or not add Mutability Flags depending on the operating system version is a hassle, not to mention adding version check. Moreover, mutability is easy to be forgotten to add. When we separate mutability into a Boolean in PendingIntentCompat, we can force this parameter be set directly at compile time to reduce runtime errors.

Test: ./gradlew test connectedCheck

This is an imported pull request from https://github.com/androidx/androidx/pull/433.

Resolves #433
Github-Pr-Head-Sha: 2206fb232e04bcee286d40a2a85e9bab454fa4fd
GitOrigin-RevId: 188a6f7bd117ee60deb2bb69425d8045c0121bad
Change-Id: Id1572e9f3aca67f8963f9166c9de64c5190795d8