[Extensions] Failed on onOutputSurface call with unassigned OutputSurface when opening the camera
After the aosp/2200238 change, onInit is invoked right before opening the capture session.
the output surface will be stored when preview output surface is ready and
CaptureProcessor.onOutputSurface is called. And this stored output surface will be set to
OEM after onInit(upon capture session creation).

The problem is that, the capture session creation only waits for the readiness of input surface.
So it is likely that the output surface is not specified upon creating the capture session.

This CL is to ensure output surface is ready and set along with input surface when ProcessingSurface
.getSurface completes.

Test: BindUnbindUseCaseStressTest,  LifecycleStatusChangeStressTest
Bug: 245260182
Change-Id: Ice901056a40fcfd9af82796e0e02e9c382f9b8a0