Created ProfileVerifier to check whether a compilation profile exists
The ProfileVerifier checks if a compilation profile is installed, either through play store, package manager or profile installer.
A reference profile can be the result of either a baseline profile, a cloud profile or a bg dex opt.
If a reference profile exists at the first app start or after an app update,
this means that a baseline profile or a cloud profile was installed, since the bg dex opt happen
later in background. The ProfileVerifier caches the result after the first app start to exclude bg dex opt
and rechecks for reference profile only after app update. The same is applied also to current profiles, with the exception
that the check for current profiles can be forced, in case the installation was done at a later time through ProfileInstallerReceiver.

Test: ./gradlew profileinstaller:integration-tests:profile-verification:cC
Bug: 246653809
Change-Id: I263a45582743afce0c3d9aefe629c07dfde77a72
Relnote: Added ProfileVerifier api to check whether a compilation profile exists for the app