Dispatch Consumers only on added fragmentManagers
The new providered callbacks are registered when the fragment manager is
attached to the host, but not unregistered until destroyed. This means
that fragments on the back stack can still have the callbacks dispatched
to their child fragment managers, even though they are not added.

aosp/2216147 already fixed this for config changes, but we need to fix
it for multi-window, picture-in-picture, and on low memory modes.

RelNote: "Fixed an error that caused fragments on the back stack to get
onMultiWindowModeChanged(), onPictureInPictureModeChanged(), and
onLowMemory() callbacks."
Test: Added ProviderCallbackTests

Change-Id: I34581e67bb3a81c7a7e162c3e47e51359896a628