Resolving AutoMigration issue when performing multiple column renames.
The issue was caused by incorrect logic in the SchemaDiffer file, particularly in detectColumnLevelChanges(...). The logic in this function was correctly identifying renamed columns, but if the renamed column was detected in a table that was already detected to be "complexly changed" (added to the complexChangedTables map), instead of updating the renamedColumnsMap of that entry in the complexChangedTables, was replacing the existing entry with a new ComplexChangedTable entry, deleting the already detected renamed columns. This has been fixed in this CL to perform this update in the case the table has already had a complex change, or to create a new entry if this is the first time a complex change was detected in the table.

Test: AutoMigrationTest testRenameTwoColumnsOnComplexChangedTable() tests particularly performing multiple renames on columns on an already complex changed table.
Bug: 244594343
Bug: 246362513
Change-Id: Icfe77f68a8a5bc30f03c3f179412299345c39a2b