Fix not emitting when resubscribing.
Fix not emitting when resubscribing on Sidecar. In the
DistinctElementSidecarCallback we were holding on to the last emitted
value. When a down stream listener would resubscribe it would not get
the value back.  The workaround would be for down stream subscribers to
maintain a copy of the last emitted value.

This does not affect devices with Extensions. We have the test
testWindowLayoutInfo_updatesOnSubsequentRegistrations to catch this
scenario.

Bug: 233764537
Test: ./gradlew window:window:cAT on a device with Sidecar.
Change-Id: I0fb409f2a549e356bafc13abea7393df576f5e3f