Improve Snapping animations
Snapping happens in two moments. Approach and Snap. In the previous implementation we were using the same animation spec
for snap and approach if the fling velocity was not large enough to decay+snap. Now we give the possibility of using a
different animation spec if the velocity is not large enough to decay + snap. We use tween + snap as the default.
Test: Updated tests to include new parameter

Relnote: Introduced lowVelocityAnimationSpec used by the approach step when there's not enough fling velocity to decay.

Change-Id: Iaeb272de77d42e142c01cc3f514a85331e9ef856