Fix memory leak in ComplicationDataSourceInfoRetriever
For reasons unknown the continuation object from
suspendCancellableCoroutine was not getting GCd and this was a
gc root that retained the activity via the compiler generated
coroutine continuations.

This patch 'fixes' this by introducing a helper class where
the callback can be explicitly nulled after we're done with it.
This is enough to prevent the activity being retained.

We also change the coroutine scope to a lifetime observer one
which automatically cancels at the appropriate times.

Test: Manual testing with android studio profiler
Bug: 249121838
Change-Id: Id3a47b91e5cca8ea4ba0c13b90d4f2dcecd4a591