Pager Demos, Samples and Publication
Make Pager APIs public and added Demos and Samples.

Relnote: Introduce HorizontalPager and VerticalPager, a way of showing composables in a Pager manner. Introduced PagerState to control the Pagers as well as query information about the Pager's current state. Introduced PageSize, a way of controlling the size of a Pager's page, this can be used to create carrousel like Pagers. Introduced PagerSnapDistance, a way to control how snapping will work in Pager's fling behavior.

Test: N/A
Change-Id: I01120224eaccd9ee255890eb409e87a7ef7ffd5f