Ensure child fragments get provider callbacks from fragmentController
If you are using a custom FragmentController instead of the one defined
by the FragmentManager, and the controller's host does not implement the
provider callback interfaces, the callbacks are not dispatched down the
fragment hierarchy, causing child fragments to fail to get any
callbacks.

We need to ensure that if a custom controller that does not implement
the callback interfaces still gets the proper callbacks, even when using
the now deprecated dispatch functions.

If the controller does implement
the interfaces, attempting to use the deprecated functions will throw an
IllegalStateException.

RelNote: "Fixed an error where using a custom FragmentController with a
host that does not implement a provider callback interface
(OnConfigurationChangedProvider, OnMultiWindowModeChangedProvider,
OnTrimMemoryProvider, OnPictureInPictureModeChangedProvider) and calling
its deprecated dispatch function would fail to dispatch to child
fragments."
Test: Added tests
Bug: 248313906

Change-Id: I9b380cd13d0fadf4a59be537d22023019f1b7cb4