Generate DAOs in Kotlin
Partially migrate DAO code generation (DaoWriter, QueryWriter) to XPoet such that the simple path of InstantQueryResultBinder -> SingleItemQueryResultAdapter -> PojoRowAdapter can be completely generated in Kotlin. With this big change we are able to start parallel migration of the different supported binders and adapters.

The following codegen classes are now completely migrated to XPoet:
* QueryWriter
* InstantQueryResultBinder
* CursorQueryResultBinder
* SingleEntityQueryResultAdapter
* PojoIndexAdapter

Note that PojoRowAdapter now in-lines the FieldReadWriteWriter generated code, but it is still partially migrated to XPoet since RelationCollectorFunctionWriter is not migrated.

New XPoet APIs:
* XMemberName - To emit companion object and top-level functions / properties references.
* XFunSpec.overridingBuilder() - To create an override function builder.
* XTypeSpec.anonymousClassBuilder() - To create anonymous classes.
* XTypeSpec.companionObjectBuilder() - To create companion object classes. For Java this is just a class named 'Companion'.

Bug: 127483380
Test: Existing, mainly the Java test app. KotlinCodeGenTest now verifies generates DAO implementations (instead of isolated PojoRowAdapter). With shared functions and properties migrated we can also now tests Enums and Custom type converters, added codegen tests for those.
Change-Id: I60033b06f1c57497fc4ddbe2678f14ba589dd566