Create compose-ui nodes for curved text elements.
For each curvedText/basicCurvedText we create a compose-ui node that has
as content description the text of the (basic)curved text, and we size
and place it appropiatelly. This enables talkback to work properly with
curved text elements.
Minor change to the internal API for CurvedChild.initializeMeasure
Test: Manual, with and withouth talkback. + ./gradlew :wear:compose:compose-foundation:connectedCheck
Relnote: "curveText and basicCurvedText will now have a properly sized
and placed (but empty) compose-ui node associated with them, using the text as
content description, to properly work with talkback"
Bug: 210721259

Change-Id: I7af7ced923180839eee782fe2e45a1976c7e8ab4