Address lint warning in PreviewActivity
The new lint warning UnusedMaterialScaffoldPaddingParameter introduced
in aosp/2017215 is triggering in this file when we try to upgrade
ui-tooling's dependency to a new version of Material.

I'm not sure it will ever happen to have nonzero padding in an Android
Studio preview so I considered suppressing the lint warning.  But
following the recommended best practice also looks harmless and in
theory more correct.

Bug: 251192712
Test: Existing tests
Change-Id: Ib04f3c2fb53bf10adcc8214ef2c152733656c4bb