Change the resizing of Wrap dimensioned Images.
After this change, appwidget Images that have a Wrap dimension as either
width or height will also take into account the source image aspect ratio
when choosing a size for the wrapped image.

This is slightly different from Compose UI, where Fill_Width and Fill_Height
are needed to match this behavior. Fill width and height require the size
of the source image to be known at layout time to decide which axis is
cropped. In AppWidget we don't get a callback for this so its not possible.

This new behavior, while slightly different from Compose UI, presents an
unsurprising result: an image that has Fill/Expand width and Wrap height,
will grow to fill its width and set its height to wrap the image, keeping
the aspect ratio.

Bug: 251549420
Test: Unit tests in ApplyDimensionModifierTest

Change-Id: Ice18c7fa0eb15e2c3ceff9829aa23bf64db16d1b