Fix removing state in onCleared()
When you call popBackStack() or hit the system
back button, the NavController will instantly update
its own state. This update triggers the NavHost to
update the visible destination, crossfading / running
transitions between the previous destination and the new
destination. This async process ends with the previous
destination leaving composition. At that point and only
at that point, can the NavController completely destroy
any state associated with that destination. This includes
clearing out any ViewModels assocaited with that
destination.

This means that the previous approach of clearing out
the reference to the SaveableStateHolder in a
DisposableEffect was too early - when the onCleared()
of our BackStackEntryIdViewModel was called, the
removal of Compose state was never able to actually
happen.

By instead keeping a WeakReference to the
SaveableStateHolder, we get two properties:

- The SaveableStateHolder is available when
onCleared() is called
- The ViewModel does not store a reference to
a SaveableStateHolder across configuration changes
(i.e., does not leak the SaveableStateHolder)

Relnote: "Fixed an issue where saved Compose state
(e.g., usages of `rememberSaveable`) were not being
forgotten and removed when a destination was popped
off the back stack."
Test: Leak regression test still passes, tested in sample
BUG: 245355393

Change-Id: I649492142419b56d9ef5ab768a2cfe7662f026e2