Migrate QueryParameterAdapter implementations to XPoet
The migrated adapters are:
 * BasicQueryParameterAdapter
 * ArrayQueryParameterAdapter
 * CollectionQueryParameterAdapter

A notable change is that nullability of the array and collection parameters have to be considered. Instead of prohibit this CL simply accepts them by binding a null arg when the param is null, essentially tackling b/72246856. This change affects all backends (javac, kapt and ksp) and Java code generation too, passing a null collection as a query param will no longer cause a crash.

Also fixed a bug in KspArrayType.asTypeName() where the KTypeName didn't had right nullability, fixed it by using the KsTypeKotlinPoetExt (it was a missed migration).

New XPoet API:
 * XCodeBlock.ofTernaryIf() - For generating inline if expressions with Java's ternary operator and Kotlin if/else.

Bug: 72246856
Bug: 127483380
Test: NullableCollectionQueryParamTest + KotlinCodeGenTest
Change-Id: I5e5b6a73733f2bbc8d111cdfa724720757ab38c5