Deprecates the top app bar's smallTopAppBarColors
The top app bar was recently renamed from SmallTopAppBar to TopAppBar.
However, the TopAppBarDefaults function for providing colors was not
renamed.
This change deprecates the smallTopAppBarColors() in favor of a new
topAppBarColors() with the same functionality.

Bug: 249476076
Test: AppBarTest passing
Relnote: "Deprecate the TopAppBarDefaults smallTopAppBarColors function
in favor of a new topAppBarColors function that should be used when
creating a TopAppBar."

Change-Id: Ie6cb94d2ff4278f1315b4acbf4d0a816afa7050f