Update CaptureSessionTest for studio-x10, itel-w6004
The request option CONTROL_CAPTURE_INTENT for testing
causes a failure on the CameraCaptureSession.

The test only required to verify the capture request
can be successfully sent to the camera, not actually
needing to set CONTROL_CAPTURE_INTENT.
In this CL, we use CONTROL_AE_EXPOSURE_COMPENSATION instead
it would be better compatiable on the devices.

Bug: 254606975
Test: CaptureSessionTest
Change-Id: Ibc07dc51db36490a6513c0a14d77ceaa387b22d3